+++ src/Atata.WebDriverExtras/Exceptions/ExceptionFactory.cs
@@ -3,14 +3,12 @@
 /// <summary>
 /// Provides a set of methods for an exception creation.
 /// </summary>
-// TODO: v4. Remove ExceptionFactory.
+// TODO: v5. Remove ExceptionFactory.
 public static class ExceptionFactory
 {
-    public static TimeoutException CreateForTimeout(TimeSpan spentTime, Exception? innerException = null)
-    {
-        string message = $"Timed out after {spentTime.TotalSeconds} seconds.";
-        return new(message, innerException);
-    }
+    [Obsolete("Use TimeoutExceptionFactory.Create(...) instead.")] // Obsolete since v4.0.0.
+    public static TimeoutException CreateForTimeout(TimeSpan spentTime, Exception? innerException = null) =>
+        TimeoutExceptionFactory.Create(spentTime, innerException);
 
     public static ArgumentException CreateForUnsupportedEnumValue<T>(
         T value,